Move ImportedLFPV1 to schema lfp_imported
