+++ src/spyglass/lfp/lfp_merge.py
@@ -4,7 +4,8 @@ import pandas as pd
 from spyglass.common.common_ephys import LFP as CommonLFP  # noqa: F401
 from spyglass.common.common_filter import FirFilterParameters  # noqa: F401
 from spyglass.common.common_interval import IntervalList  # noqa: F401
-from spyglass.lfp.v1.lfp import LFPV1, ImportedLFPV1  # noqa: F401
+from spyglass.lfp.v1.lfp import LFPV1  # noqa: F401
+from spyglass.lfp.lfp_imported import ImportedLFPV1  # noqa: F401
 from spyglass.utils.dj_merge_tables import _Merge
 
 schema = dj.schema("lfp_merge")
