+++ src/spyglass/lfp/v1/lfp.py
@@ -176,21 +176,3 @@ class LFPV1(dj.Computed):
             nwb_lfp["lfp"].data,
             index=pd.Index(nwb_lfp["lfp"].timestamps, name="time"),
         )
-
-
-@schema
-class ImportedLFPV1(dj.Imported):
-    definition = """
-    -> Session                      # the session to which this LFP belongs
-    -> LFPElectrodeGroup            # the group of electrodes to be filtered
-    -> IntervalList # the original set of times to be filtered
-    lfp_object_id: varchar(40)      # the NWB object ID for loading this object from the file
-    ---
-    lfp_sampling_rate: float        # the sampling rate, in samples/sec
-    -> AnalysisNwbfile
-    """
-
-    def make(self, key):
-        raise NotImplementedError(
-            "For `insert`, use `allow_direct_insert=True`"
-        )
